refactoring and first draft of actuate method
